requirements.txt
fastapi 
uvicorn 
Pillow 
requests 
python-dotenv 
ffmpeg-python 
numpy 
pandas
psutil 
cached_property 
einops  
langchain 
langchain-openai 
langchain-core 
langchain-google-genai
validators 
qdrant-client 
pillow-heif
fastapi-key-auth
black
pyopenssl
flake8
coverage